Hard limit of 35kb on file uploads

+++ routes/corpora.js
@@ -56,29 +56,41 @@ router.get('/:id', function (req, res) {
 });
 
 router.post('', function(req, res) {
- if (!req.user) {
+  if (!req.user) {
     res.status(401).json({
       message: "Unauthorized",
       error: 401
     });
   } else {
     var file = req.files.file;
-    fs.readFile(file.path, function(err, data){
-      var corpus = {
-        user_id: req.user._id,
-        contents: data,
-        title: req.body.title,
-        fileSize: file.size,
-        fileName: file.name,
-        fileType: file.type
-      };
+    var size = fs.statSync(file.path)['size'];
 
-      Corpus.create(corpus, function(err, c) {
-        res.status(201).json(c);
+    if(size <= 35000){
+
+      fs.readFile(file.path, function(err, data){
+        var corpus = {
+          user_id: req.user._id,
+          contents: data,
+          title: req.body.title,
+          fileSize: file.size,
+          fileName: file.name,
+          fileType: file.type
+        };
+
+        Corpus.create(corpus, function(err, c) {
+          res.status(201).json(c);
+        });
+
+        fs.unlink(file.path);
       });
+    }
 
-      fs.unlink(file.path);
-    });
+    else {
+      res.status(413).json({
+        message: "File size is too large!",
+        error: 413
+      });
+    }
   }
 });
